change cache interface to make it more flexible
